==> usrightmedia/code/02-twitter/02-twitter-prep-inca.ipynb <==
# 02-twitter-prep-inca
**Purpose**:  collect politicians' timelines on Twitter.

import os

from inca import Inca

dir_inp = os.path.join('..', '..', 'data', '02-intermediate', '01-congress-legislators')
import pandas as pd
from usrightmedia.shared.datetime_utils import (END_TIME, END_TIME_TEST,
                                             START_TIME, START_TIME_TEST)

from usrightmedia.shared.loggers import get_logger
LOGGER = get_logger(filename = '02-twitter-prep-inca', logger_type='main')

# list of politicians' Twitter screen_names
screen_names =  list(pd.read_pickle(os.path.join(dir_inp, 'politicians-republican.pkl'))['twitter'])
# remove None values (i.e. politician doesn't have a Twitter account)
screen_names = sorted([p for p in screen_names if p])
print(f"number of Twitter screen_names to scrape: {len(screen_names)}")

screen_names = screen_names
print(screen_names)

myinca = Inca()
LOGGER.info(f"{myinca.database.list_apps()}") # check that the 'usrightmedia' app exists
# myinca.database.delete_doctype('tweets2')

def retrieve_user_timeline(screen_name, app_name='usrightmedia'):
    """retrieve tweets from specified start_time to end_time

    Args:
        screen_name (str): Twitter screen_name
        app_name (str): name of Twitter v2 app in INCA

    Returns:
        None
        - each tweet is stored as a doc in INCA's Elasticsearch database
        - doc['_source'] contains the tweet payload
        - the doc's doctype in ES is 'tweets2'
    
    """
    myinca.clients.twitter2_timeline(
        app=app_name,
        screen_name=screen_name,
        start_time=START_TIME,
        end_time=END_TIME
    )

LOGGER.info(f"Start collection of Twitter account timelines")

for n, screen_name in enumerate(screen_names):
    count = n + 1
    LOGGER.info(f"Collecting {count} of {len(screen_names)}: {screen_name}")
    retrieve_user_timeline(screen_name)

LOGGER.info(f"Completed collection of Twitter account timelines")
